+++ src/shared/styles/userList.js
@@ -1,29 +1,8 @@
 import { defaultColor } from "~/shared/styles/colors";
 
 export default {
-    wideBody: {
-        top: "40px",
-        height: "calc(100% - 40px)",
+    center: {
         textAlign: "center",
-    },
-    centerer: {
-        textAlign: "center",
-    },
-    teamInfo: {
-        marginLeft: "15px",
-        marginLeft: "15px",
-        span: {
-            display: "inline-block",
-            verticalAlign: "textTop",
-            marginTop: "40px",
-            marginBottom: "0px",
-            marginLeft: "20px",
-            marginRight: "20px",
-        },
-        h1: {
-            paddingTop: "50px",
-            marginTop: "0px",
-        },
     },
     memberList: {
         marginTop: "20px",
