+++ src/team/styles.js
@@ -1,0 +1,23 @@
+export default {
+    wideBody: {
+        top: "40px",
+        height: "calc(100% - 40px)",
+        textAlign: "center",
+    },
+    teamInfo: {
+        marginLeft: "15px",
+        marginLeft: "15px",
+        span: {
+            display: "inline-block",
+            verticalAlign: "textTop",
+            marginTop: "40px",
+            marginBottom: "0px",
+            marginLeft: "20px",
+            marginRight: "20px",
+        },
+        h1: {
+            paddingTop: "50px",
+            marginTop: "0px",
+        },
+    },
+}
